Rename stream_reader.hpp and update related files.
