+++ src/stream.hpp
